data-converter/requirements.txt
beautifulsoup4==4.12.3
Flask==2.3.2
openai==1.12.0
PyGithub==2.2.0
PyGithub==2.2.0
python-dotenv==1.0.1
Requests==2.31.0
